feat(assessment): implement quiz and question management features
Changes:
- Added new API endpoints for creating quizzes and importing questions from AI.
- Introduced data transfer objects (DTOs) for quiz and question requests/responses.
- Implemented services for handling quiz creation and question import logic.
- Updated database schema to support new features, including migration scripts.
- Enhanced orchestration to route requests for quizzes and questions through the assessment service.

+++ src/Services/Assessment/Beyond8.Assessment.Domain/Entities/Question.cs
@@ -22,7 +22,7 @@ public class Question : BaseEntity
 
     public DifficultyLevel Difficulty { get; set; } = DifficultyLevel.Medium;
 
-    public int Points { get; set; } = 1;
+    public decimal Points { get; set; } = 1.0m;
 
     [Column(TypeName = "jsonb")]
     public string Tags { get; set; } = "[]";
